Move `Key` from `db::types` to `model`
