+++ backend/src/api/id.rs
@@ -4,7 +4,7 @@ use serde::{Deserialize, Serialize};
 use static_assertions::const_assert;
 use std::fmt;
 
-use crate::{db::types::Key, util::{base64_decode, BASE64_DIGITS}};
+use crate::model::Key;
 
 
 /// An opaque, globally-unique identifier for all "nodes" that the GraphQL API
@@ -138,30 +138,6 @@ impl Id {
     }
 }
 
-impl Key {
-    pub(crate) fn from_base64(s: &str) -> Option<Self> {
-        if s.len() != 11 {
-            return None;
-        }
-
-        decode_base64(s.as_bytes())
-    }
-
-    pub(crate) fn to_base64<'a>(&self, out: &'a mut [u8; 11]) -> &'a str {
-        // Base64 encoding. After this loop, `n` is always 0, because `u64::MAX`
-        // divided by 64 eleven times is 0.
-        let mut n = self.0;
-        for i in (0..out.len()).rev() {
-            out[i] = BASE64_DIGITS[(n % 64) as usize];
-            n /= 64;
-        }
-        debug_assert!(n == 0);
-
-        std::str::from_utf8(out)
-            .expect("bug: base64 did produce non-ASCII character")
-    }
-}
-
 impl std::str::FromStr for Id {
     type Err = &'static str;
 
@@ -191,31 +167,10 @@ impl fmt::Display for Id {
     }
 }
 
-
-fn decode_base64(src: &[u8]) -> Option<Key> {
-    let src: [u8; 11] = src.try_into().ok()?;
-
-    // Make sure the string doesn't decode to a number > `u64::MAX`. Luckily,
-    // checking that is easy. `u64::MAX` encodes to `P__________`, so the next
-    // higher number would carry through and make the highest digit a `Q`. So we
-    // just make sure the first digit is between 'A' and 'P'.
-    if src[0] > b'P' || src[0] < b'A' {
-        return None;
-    }
-
-    src.iter()
-        .rev()
-        .enumerate()
-        .map(|(i, &d)| base64_decode(d).map(|n| n as u64 * 64u64.pow(i as u32)))
-        .sum::<Option<u64>>()
-        .map(Key)
-}
-
-
 #[cfg(test)]
 mod tests {
     use std::str::FromStr;
-    use super::{Id, Key, BASE64_DIGITS};
+    use super::{Id, Key};
 
     #[test]
     fn simple() {
@@ -265,7 +220,7 @@ mod tests {
                 let id = Id { kind: Id::REALM_KIND, key: Key((n as u64) << shift) };
                 let s = id.to_string();
                 assert_eq!(s[..2].as_bytes(), Id::REALM_KIND);
-                assert!(s[2..].bytes().all(|d| BASE64_DIGITS.contains(&d)));
+                assert!(s[2..].bytes().all(|d| crate::util::BASE64_DIGITS.contains(&d)));
             }
         }
     }
