+++ backend/src/model/mod.rs
@@ -7,8 +7,10 @@
 //! either of `db`, `api` or any other submodule as they are used in multiple
 //! situations (loading from DB, exposing via API, ...).
 
+mod key;
 mod translated_string;
 
 pub(crate) use self::{
+    key::Key,
     translated_string::{LangKey, TranslatedString},
 };
