+++ backend/src/rss.rs
@@ -6,8 +6,9 @@ use futures::TryStreamExt;
 use ogrim::xml;
 
 use crate::{
-    db::{types::{EventTrack, Key}, self, util::{impl_from_db, FromDb, dbargs}},
+    db::{types::EventTrack, self, util::{impl_from_db, FromDb, dbargs}},
     http::{Context, response::{bad_request, self, not_found}, Response},
+    model::Key,
     config::HttpHost,
     prelude::*,
 };
