+++ backend/src/search/playlist.rs
@@ -5,7 +5,8 @@ use tokio_postgres::GenericClient;
 
 use crate::{
     prelude::*,
-    db::{types::Key, util::{collect_rows_mapped, impl_from_db}},
+    model::Key,
+    db::util::{collect_rows_mapped, impl_from_db},
 };
 
 use super::{realm::Realm, util::{self, FieldAbilities}, IndexItem, IndexItemKind, SearchId};
