+++ backend/src/search/series.rs
@@ -5,7 +5,8 @@ use tokio_postgres::GenericClient;
 
 use crate::{
     prelude::*,
-    db::{types::{Key, SearchThumbnailInfo}, util::{collect_rows_mapped, impl_from_db}},
+    model::Key,
+    db::{types::SearchThumbnailInfo, util::{collect_rows_mapped, impl_from_db}},
 };
 
 use super::{realm::Realm, util::{self, FieldAbilities}, IndexItem, IndexItemKind, SearchId};
